+++ docs/conf.py
@@ -24,7 +24,6 @@
 
 import os
 
-
 html_show_sourcelink = True
 nbsphinx_prolog = """`Download this Jupyter notebook <https://raw.githubusercontent.com/scipp/scipp/master/docs/{{ env.doc2path(env.docname, base=None) }}>`_
 
@@ -41,14 +40,12 @@ nbsphinx_prolog = """`Download this Jupyter notebook <https://raw.githubusercont
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.autosummary',
-    'sphinx.ext.intersphinx',
-    'sphinx.ext.mathjax',
-    'IPython.sphinxext.ipython_directive',
-    'IPython.sphinxext.ipython_console_highlighting',
-    'nbsphinx'
+    'sphinx.ext.autodoc', 'sphinx.ext.autosummary', 'sphinx.ext.intersphinx',
+    'sphinx.ext.mathjax', 'IPython.sphinxext.ipython_directive',
+    'IPython.sphinxext.ipython_console_highlighting', 'nbsphinx'
 ]
+
+autodoc_typehints = 'description'
 
 intersphinx_mapping = {
     'numpy': ('https://docs.scipy.org/doc/numpy/', None),
@@ -104,7 +101,6 @@ pygments_style = 'sphinx'
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 todo_include_todos = False
 
-
 # -- Options for HTML output ----------------------------------------------
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
@@ -118,11 +114,7 @@ if not on_rtd:  # only import and set the theme if we're building docs locally
     html_theme = 'sphinx_rtd_theme'
     html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 
-    html_context = {
-        'css_files': [
-            '_static/theme_overrides.css'
-        ]
-    }
+    html_context = {'css_files': ['_static/theme_overrides.css']}
 else:
     html_context = {
         'css_files': [
@@ -136,9 +128,7 @@ else:
 # further.  For a list of options available for each theme, see the
 # documentation.
 #
-html_theme_options = {
-    'logo_only': True
-}
+html_theme_options = {'logo_only': True}
 
 html_logo = "_static/logo-large-v4.png"
 # html_favicon = "_static/favicon.ico"
@@ -161,12 +151,10 @@ html_sidebars = {
     ]
 }
 
-
 # -- Options for HTMLHelp output ------------------------------------------
 
 # Output file base name for HTML help builder.
 htmlhelp_basename = 'scippdoc'
-
 
 # -- Options for LaTeX output ---------------------------------------------
 
@@ -192,20 +180,15 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'scipp.tex', u'scipp Documentation',
-     u'Simon Heybrock', 'manual'),
+    (master_doc, 'scipp.tex', u'scipp Documentation', u'Simon Heybrock',
+     'manual'),
 ]
-
 
 # -- Options for manual page output ---------------------------------------
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [
-    (master_doc, 'scipp', u'scipp Documentation',
-     [author], 1)
-]
-
+man_pages = [(master_doc, 'scipp', u'scipp Documentation', [author], 1)]
 
 # -- Options for Texinfo output -------------------------------------------
 
@@ -213,11 +196,9 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'scipp', u'scipp Documentation',
-     author, 'scipp', 'One line description of project.',
-     'Miscellaneous'),
+    (master_doc, 'scipp', u'scipp Documentation', author, 'scipp',
+     'One line description of project.', 'Miscellaneous'),
 ]
-
 
 # -- Options for Matplotlib in notebooks ----------------------------------