+++ python/tests/plot/test_plot_nd.py
@@ -38,6 +38,10 @@ def test_plot_sliceviewer_2d_with_labels():
     plot(make_dense_dataset(ndim=3, labels=True), axes={'x': "somelabels"})
 
 
+def test_plot_sliceviewer_2d_with_attrs():
+    plot(make_dense_dataset(ndim=3, attrs=True), axes={'x': "attr"})
+
+
 def test_plot_sliceviewer_2d_with_binedges():
     plot(make_dense_dataset(ndim=3, binedges=True))
 
@@ -57,7 +61,7 @@ def test_plot_variable_3d():
     plot(v3d)
 
 
-def test_plot_4d_with_masks():
+def test_plot_4d_with_masks_no_coords():
     data = sc.DataArray(data=sc.Variable(
         dims=['pack', 'tube', 'straw', 'pixel'],
         values=np.random.rand(2, 8, 7, 256)),
@@ -68,6 +72,7 @@ def test_plot_4d_with_masks():
                                           values=np.where(
                                               a > 0.5, True, False))
     plot(data)
+    plot(data, axes={'y': 'tube'})
 
 
 def test_plot_3d_data_ragged():